balance list items update UI

+++ feature-wallet-impl/src/main/java/jp/co/soramitsu/feature_wallet_impl/presentation/balance/list/model/BalanceModel.kt
@@ -2,13 +2,12 @@ package jp.co.soramitsu.feature_wallet_impl.presentation.balance.list.model
 
 import jp.co.soramitsu.feature_wallet_impl.presentation.model.AssetModel
 import java.math.BigDecimal
-import java.math.BigInteger
 
 class BalanceModel(val assetModels: List<AssetModel>) {
     val totalBalance = calculateTotalBalance()
 
     private fun calculateTotalBalance(): BigDecimal {
-        return assetModels.fold(BigDecimal(BigInteger.ZERO)) { acc, current ->
+        return assetModels.fold(BigDecimal.ZERO) { acc, current ->
             val toAdd = current.dollarAmount ?: BigDecimal.ZERO
 
             acc + toAdd
